Fix cut out content at bottom of screen
The height of the main section was set by subtracting the fixed
height of the header from the height of the screen (100vh). The
sidebar and content area were grouped so that this height was only
applied to their parent. Doing this also made each section
scrollable independently withouth having to use absolute
positioning.